requirements.txt
Flask==2.2.3
werkzeug==2.2.3
moviepy==2.1.2
Pillow>=9.2.0,<11.0
imageio==2.22.4
numpy>=1.25.0
proglog==0.1.10
python-dotenv==0.21.0
